Add files via upload
post jawha behy

+++ src/main/java/tn/esprit/payment/models/Payment.kt
@@ -1,11 +1,12 @@
 package tn.esprit.payment.models
- data class Payment(
-        val amount: Double,
-        val date: String,
-        val method: String,
-        val numberOfRoommates: Int,
-        val isRecurringPayment: Boolean,
-        val recurringPaymentFrequency: String,
-        val paymentType: List<String>
+
+data class Payment(
+    val amount: Int,
+    val date: String,
+    val method: String,
+    val numberOfRoommates: Int,
+    val isRecurringPayment: Boolean,
+    val recurringPaymentFrequency: String,
+    val paymentType: List<String>
     )
